revenue/requirements.txt
flask==2.0.1
werkzeug>=2.0,<3.0
numpy==1.21.6
pandas==1.3.5
tensorflow==2.10.0
scikit-learn==1.0.2
psycopg2-binary==2.9.5
python-dateutil==2.8.2
sqlalchemy==1.4.41
gunicorn
gevent
